Add repeatability into sequencing
This commit overhauls sequencing and the lifetime of sheets/sequences, 
by providing a means to cleanly repeat a given sheet after a period.

It also cleans up lifetime management by fully initializing sheets at 
the start of a run, and makes calls to setup() and teardown() 
non-destructive, which is important for nested sheet usage within a 
repeating sheet.